proper dry wet + pull out noise to extra file
